fix: 코테, DML_SELECT
fix: 코테, DML_SELECT

+++ javaBasic/src/codingTestChallenge/Day01.java
@@ -8,8 +8,8 @@ public class Day01 {
 		
 		Scanner sc = new Scanner(System.in);		
 		long a = sc.nextLong();
-		long b = sc.nextInt();
-		long c = sc.nextInt();
+		long b = sc.nextLong();
+		long c = sc.nextLong();
 		
 		System.out.println(a+b+c);
